move scn2img related code into scn2img.py
scn2img needs some functions,variables,etc. from webui.py.
to avoid circular import, the scn2img function is wrapped
with get_scn2img. it takes the necessary variables as arguments
and returns a usable scn2img function.

monocular depth estimation model loading is now delayed until it
is actually used by used scn2img.
